+++ web/book/highlight-prql.js
@@ -61,25 +61,8 @@ formatting = function (hljs) {
     "rank",
     "rank_dense",
     "row_number",
-    // Mathematical-functions
-    "abs",
-    "floor",
-    "ceil",
-    "pi",
-    "exp",
-    "ln",
-    "log10",
-    "log",
-    "sqrt",
-    "degrees",
-    "radians",
-    "cos",
-    "acos",
-    "sin",
-    "asin",
-    "tan",
-    "atan",
-    "pow",
+    // Mathematical-functions module
+    "math",
   ];
 
   const DATATYPES = [